requirements.txt
fastapi
uvicorn[standard]
httpx
tiktoken
sentence-transformers
faiss-cpu
numpy
